+++ server.js
@@ -17,6 +17,7 @@ app.use(bodyParser.json());
 
 // Routes
 const users = require('./routes/users');
+const friends = require('./routes/friends');
 
 // DB config
 const db = require('./config/db').mongoURI;
@@ -27,6 +28,7 @@ mongoose
     .catch(err => console.log(err));
 
 app.use('/users', users);
+app.use('/friends', friends);
 
 const PORT = process.env.PORT || 3001;
